fix: Hover is not reset properly with touchscreens.
When hoverEnabled:true is set, the hover state is still true when
the button is released under the touch screen. So the hover state
should be reset when the QEvent::TouchEnd event is detected.

In addition, long press the button under the touch screen and release
it, it will trigger the QEvent::hoverMove event, which will also cause
hover is true after release. Reading the QEvent::hoverMove event
documentation, it doesn't make sense for Button.
Therefore, the QEvent::hoverMove event is not handled in the button.

Bug: https://pms.uniontech.com/bug-view-155535.html

log: Fix the abnormal state of button hover on the touch screen.